requirements.txt
langchain_community
langchain_experimental 
matplotlib 
langgraph
python-dotenv
google-generativeai
langchain-google-genai

